refactor: split UI into components and consolidate data logic into a custom hook
